bridge/requirements.txt
paho-mqtt
kafka-python
backoff